+++ CHANGELOG.md
@@ -10,6 +10,7 @@
 - Capabilities are now resolved using `CapabilitiesResolver` class.
 - Require PHPUnit ^5.7
 - Remove dependency on unmaintained Configula (and internally reimplement configuration options retrieval)
+- Improve `install` command output (eg. to always include path to downloaded file).
 
 ## 2.1.0 - 2017-01-16
 ### Added